requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.8.0
requests>=2.31.0
